Split ReadQuery adapters into their own modules
