+++ tiledb/api/src/query/read/raw.rs
@@ -1,0 +1,246 @@
+use super::*;
+
+pub(crate) struct RawReadOutput<'data, C> {
+    pub data_size: Pin<Box<u64>>,
+    pub offsets_size: Option<Pin<Box<u64>>>,
+    pub location: &'data RefCell<OutputLocation<'data, C>>,
+    pub raw_dataptr: *const C,
+    pub raw_offsetptr: Option<*const u64>,
+}
+
+impl<'data, C> RawReadOutput<'data, C> {
+    pub fn new(location: &'data RefCell<OutputLocation<'data, C>>) -> Self {
+        let (data, cell_offsets) = {
+            let mut scratch: RefMut<OutputLocation<'data, C>> =
+                location.borrow_mut();
+
+            let data = scratch.data.as_mut() as *mut [C];
+            let data = unsafe { &mut *data as &mut [C] };
+
+            let cell_offsets = scratch.cell_offsets.as_mut().map(|c| {
+                let c = c.as_mut() as *mut [u64];
+                unsafe { &mut *c as &mut [u64] }
+            });
+
+            (data, cell_offsets)
+        };
+
+        let (data_size, offsets_size) = {
+            (
+                Box::pin(std::mem::size_of_val(&*data) as u64),
+                cell_offsets.as_ref().map(|off| {
+                    let sz = std::mem::size_of_val::<[u64]>(&*off);
+                    Box::pin(sz as u64)
+                }),
+            )
+        };
+
+        let (raw_dataptr, raw_offsetptr) = {
+            let location = location.borrow();
+            (
+                location.data.as_ptr(),
+                location.cell_offsets.as_ref().map(|c| c.as_ptr()),
+            )
+        };
+
+        RawReadOutput {
+            data_size,
+            offsets_size,
+            location,
+            raw_dataptr,
+            raw_offsetptr,
+        }
+    }
+
+    fn attach_query<S>(
+        &mut self,
+        context: &Context,
+        c_query: *mut ffi::tiledb_query_t,
+        field: &S,
+    ) -> TileDBResult<()>
+    where
+        S: AsRef<str>,
+    {
+        let c_context = context.capi();
+        let c_name = cstring!(field.as_ref());
+
+        let mut location = self.location.borrow_mut();
+
+        *self.data_size.as_mut() =
+            std::mem::size_of_val::<[C]>(&location.data) as u64;
+
+        context.capi_return({
+            let data = &mut location.data;
+            let c_bufptr = data.as_mut().as_ptr() as *mut std::ffi::c_void;
+            let c_sizeptr = self.data_size.as_mut().get_mut() as *mut u64;
+
+            unsafe {
+                ffi::tiledb_query_set_data_buffer(
+                    c_context,
+                    c_query,
+                    c_name.as_ptr(),
+                    c_bufptr,
+                    c_sizeptr,
+                )
+            }
+        })?;
+
+        let cell_offsets = &mut location.cell_offsets;
+
+        if let Some(ref mut offsets_size) = self.offsets_size.as_mut() {
+            let cell_offsets = cell_offsets.as_mut().unwrap();
+
+            *offsets_size.as_mut() =
+                std::mem::size_of_val::<[u64]>(cell_offsets) as u64;
+
+            let c_offptr = cell_offsets.as_mut_ptr() as *mut u64;
+            let c_sizeptr = offsets_size.as_mut().get_mut() as *mut u64;
+
+            context.capi_return(unsafe {
+                ffi::tiledb_query_set_offsets_buffer(
+                    c_context,
+                    c_query,
+                    c_name.as_ptr(),
+                    c_offptr,
+                    c_sizeptr,
+                )
+            })?;
+        }
+
+        Ok(())
+    }
+}
+
+pub struct RawReadQuery<'data, C, Q> {
+    pub(crate) field: String,
+    pub(crate) raw_read_output: RawReadOutput<'data, C>,
+    pub(crate) base: Q,
+}
+
+impl<'ctx, 'data, C, Q> ContextBound<'ctx> for RawReadQuery<'data, C, Q>
+where
+    Q: ContextBound<'ctx>,
+{
+    fn context(&self) -> &'ctx Context {
+        self.base.context()
+    }
+}
+
+impl<'data, C, Q> QueryCAPIInterface for RawReadQuery<'data, C, Q>
+where
+    Q: QueryCAPIInterface,
+{
+    fn raw(&self) -> &RawQuery {
+        self.base.raw()
+    }
+}
+
+impl<'ctx, 'data, C, Q> ReadQuery<'ctx> for RawReadQuery<'data, C, Q>
+where
+    Q: ReadQuery<'ctx>,
+{
+    type Intermediate = (usize, usize, Q::Intermediate);
+    type Final = (usize, usize, Q::Final);
+
+    fn step(
+        &mut self,
+    ) -> TileDBResult<ReadStepOutput<Self::Intermediate, Self::Final>> {
+        /* update the internal buffers */
+        self.raw_read_output.attach_query(
+            self.context(),
+            **self.raw(),
+            &self.field,
+        )?;
+
+        /* then execute */
+        let base_result = {
+            let _ = self.raw_read_output.location.borrow_mut();
+            self.base.step()?
+        };
+
+        let records_written = match self.raw_read_output.offsets_size.as_ref() {
+            Some(offsets_size) => {
+                **offsets_size as usize / std::mem::size_of::<u64>()
+            }
+            None => {
+                *self.raw_read_output.data_size as usize
+                    / std::mem::size_of::<C>()
+            }
+        };
+        let bytes_written = *self.raw_read_output.data_size as usize;
+
+        Ok(match base_result {
+            ReadStepOutput::NotEnoughSpace => {
+                /* TODO: check that records/bytes are zero and produce an internal error if not */
+                ReadStepOutput::NotEnoughSpace
+            }
+            ReadStepOutput::Intermediate(base_result) => {
+                if records_written == 0 && bytes_written == 0 {
+                    ReadStepOutput::NotEnoughSpace
+                } else if records_written == 0 || bytes_written == 0 {
+                    /* TODO: internal error */
+                    return Err(unimplemented!());
+                } else {
+                    ReadStepOutput::Intermediate((
+                        records_written,
+                        bytes_written,
+                        base_result,
+                    ))
+                }
+            }
+            ReadStepOutput::Final(base_result) => ReadStepOutput::Final((
+                records_written,
+                bytes_written,
+                base_result,
+            )),
+        })
+    }
+}
+
+pub struct RawReadBuilder<'data, C, B> {
+    pub(crate) field: String,
+    pub(crate) raw_read_output: RawReadOutput<'data, C>,
+    pub(crate) base: B,
+}
+
+impl<'ctx, 'data, C, B> ContextBound<'ctx> for RawReadBuilder<'data, C, B>
+where
+    B: ContextBound<'ctx>,
+{
+    fn context(&self) -> &'ctx Context {
+        self.base.context()
+    }
+}
+
+impl<'data, C, B> QueryCAPIInterface for RawReadBuilder<'data, C, B>
+where
+    B: QueryCAPIInterface,
+{
+    fn raw(&self) -> &RawQuery {
+        self.base.raw()
+    }
+}
+
+impl<'ctx, 'data, C, B> QueryBuilder<'ctx> for RawReadBuilder<'data, C, B>
+where
+    B: QueryBuilder<'ctx>,
+{
+    type Query = RawReadQuery<'data, C, B::Query>;
+
+    fn array(&self) -> &Array {
+        self.base.array()
+    }
+
+    fn build(self) -> Self::Query {
+        RawReadQuery {
+            field: self.field,
+            raw_read_output: self.raw_read_output,
+            base: self.base.build(),
+        }
+    }
+}
+
+impl<'ctx, 'data, C, B> ReadQueryBuilder<'ctx> for RawReadBuilder<'data, C, B> where
+    B: ReadQueryBuilder<'ctx>
+{
+}
